Part 3: Cleans up the SQLite factories into another file/extension
