+++ Storage/SQL/SQLiteHistoryRecommendations.swift
@@ -28,7 +28,6 @@ private enum BookmarkSource {
 }
 
 extension SQLiteHistory: HistoryRecommendations {
-
     public func getHighlights() -> Deferred<Maybe<Cursor<Site>>> {
         let limit = 19
         let microsecondsPerMinute: Timestamp = 60_000_000 // 1000 * 1000 * 60
@@ -94,49 +93,5 @@ extension SQLiteHistory: HistoryRecommendations {
             "AND \(source.tableName).type = 1",
             "GROUP BY history.url",
         ].joinWithSeparator(" ")
-    }
-
-    private class func iconHistoryColumnFactory(row: SDRow) -> Site {
-        let site = basicHistoryColumnFactory(row)
-        site.icon = iconColumnFactory(row)
-        return site
-    }
-
-    private class func basicHistoryColumnFactory(row: SDRow) -> Site {
-        let id = row["historyID"] as! Int
-        let url = row["url"] as! String
-        let title = row["title"] as! String
-        let guid = row["guid"] as! String
-
-        // Extract a boolean from the row if it's present.
-        let iB = row["is_bookmarked"] as? Int
-        let isBookmarked: Bool? = (iB == nil) ? nil : (iB! != 0)
-
-        let site = Site(url: url, title: title, bookmarked: isBookmarked)
-        site.guid = guid
-        site.id = id
-
-        // Find the most recent visit, regardless of which column it might be in.
-        let local = row.getTimestamp("localVisitDate") ?? 0
-        let remote = row.getTimestamp("remoteVisitDate") ?? 0
-        let either = row.getTimestamp("visitDate") ?? 0
-
-        let latest = max(local, remote, either)
-        if latest > 0 {
-            site.latestVisit = Visit(date: latest, type: VisitType.Unknown)
-        }
-
-        return site
-    }
-
-    private class func iconColumnFactory(row: SDRow) -> Favicon? {
-        if let iconType = row["iconType"] as? Int,
-            let iconURL = row["iconURL"] as? String,
-            let iconDate = row["iconDate"] as? Double,
-            let _ = row["iconID"] as? Int {
-            let date = NSDate(timeIntervalSince1970: iconDate)
-            return Favicon(url: iconURL, date: date, type: IconType(rawValue: iconType)!)
-        }
-        return nil
     }
 }
